move Deps struct to tasks.rs
